fix: rename error
